+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.5.2
 * Bugfix for copy-and-edit paste being stuck in loading limbo.
+* Bugfix for refreshing pastes not working properly.
 
 ## 1.5.1
 * Added webpack config for splitChunks. Prod was using only a single chunk.